Avoid re-storing query and index when clicking more in search results page.

+++ peerpalette.com/SearchPage.py
@@ -11,7 +11,7 @@ import random
 
 class SearchPage(RequestHandler):
   def get(self):
-    q = self.request.get('q')
+    q = self.request.get('q').strip()
 
     if not q:
       self.redirect("/")
@@ -39,6 +39,7 @@ class SearchPage(RequestHandler):
     existing_chats = {}
     users_status = {}
     unread_chats = {}
+    cursor = None
 
     res_keys = search_query.fetch(config.ITEMS_PER_PAGE)
     if len(res_keys) >= config.ITEMS_PER_PAGE:
@@ -79,36 +80,36 @@ class SearchPage(RequestHandler):
 
       result_values.append(v)
 
-    context_text = ""
-    if context:
-      context_text = context.context
-      keyword_hashes = list(keyword_hashes + search.get_keyword_hashes(search.clean_query_string(context_text)))[:config.MAX_KEYWORDS]
+    if not cursor:
+      context_text = ""
+      if context:
+        context_text = context.context
+        keyword_hashes = list(keyword_hashes + search.get_keyword_hashes(search.clean_query_string(context_text)))[:config.MAX_KEYWORDS]
+      query = existing_query.get_result()
+      if query is not None:
+        existing_index = db.Query(models.QueryIndex, keys_only = True).filter('query =', query).get()
+        if existing_index is not None:
+          db.delete(existing_index)
+        query.query_string = q
+        query.context = context_text
+        query.date_time = self.now
+      else:
+        query = models.Query(key = query_key, query_string = q, context = context_text, date_time = self.now)
 
-    query = existing_query.get_result()
-    if query is not None:
-      existing_index = db.Query(models.QueryIndex, keys_only = True).filter('query =', query).get()
-      if existing_index is not None:
-        db.delete(existing_index)
-      query.query_string = q
-      query.context = context_text
-      query.date_time = self.now
-    else:
-      query = models.Query(key = query_key, query_string = q, context = context_text, date_time = self.now)
+      index = models.QueryIndex(key_name = search.encode_query_index_key_name(query_key), query = query_key, user = self.user_key, keyword_hashes = keyword_hashes)
+      if online_count < 5:
+        top_search = models.TopSearch(key_name = query_key.name(), query_string = q, rating = float(len(search_hashes) * online_count))
+        db.put([query, index, top_search])
+      else:
+        db.put([query, index])
 
     if not result_values:
       self.template_values['top_searches'] = get_top_searches()
 
-    index = models.QueryIndex(key_name = search.encode_query_index_key_name(query_key), query = query_key, user = self.user_key, keyword_hashes = keyword_hashes)
-    top_search = models.TopSearch(key_name = query_key.name(), query_string = q, rating = float(len(search_hashes) * online_count))
-    db.put([query, index, top_search])
-
     self.template_values["results"] = result_values
     self.template_values["key"] = query.key()
     self.template_values["query"] = q
-    try:
-      self.template_values["cursor"] = cursor
-    except NameError:
-      pass
+    self.template_values["cursor"] = cursor
 
     self.render_page('SearchPage.html')
